+++ app/Main.hs
@@ -2,7 +2,7 @@ module Main where
 
 import Expression
 import Derivative
-import Pretty
+import Simplify
 
 main :: IO ()
 main = putStrLn "Nothing to see here, yet"