+++ src/Expression.hs
@@ -18,13 +18,13 @@ data MathExpr a =
   | Coef a
   | Func1 UnaryOp (MathExpr a)
   | Func2 BinOp (MathExpr a) (MathExpr a)
-  deriving (Eq,Show,Read)
+  deriving (Eq)
 
 data BinOp = Add | Mult 
-  deriving (Show,Eq,Read)
+  deriving (Eq)
 
 data UnaryOp = Cos | Sin | Abs | Power Int 
-  deriving (Show,Eq,Read)
+  deriving (Eq)
 
 
 {-
@@ -71,3 +71,27 @@ instance Floating a => Floating (MathExpr a) where
   atanh _ = error "atanh is left un-implemented for now"
   exp _   = error "exp is left un-implemented for now"
   sqrt _  = error "sqrt is left un-implemented for now"
+
+
+
+
+
+{-
+  ======================================================================
+  implements the Show type class for MathExpr
+  ======================================================================
+-}
+instance Show a => Show (MathExpr a) where
+  show X             = "X"
+  show (Coef a)      = show a
+  show (Func1 op ex) =
+    case op of
+      Cos     -> "cos(" ++ show ex ++ ")"
+      Sin     -> "sin(" ++ show ex ++ ")"
+      Abs     -> "|" ++ show ex ++ "|"
+      Power n -> "(" ++ show ex ++ ")^(" ++ show n ++ ")"
+  show (Func2 op ex1 ex2) =
+    case op of
+      Add     -> "(" ++ show ex1 ++ ") + (" ++ show ex2 ++ ")"
+      Mult    -> "(" ++ show ex1 ++ ") * (" ++ show ex2 ++ ")"
+